+++ server/db/schema/messages.ts
@@ -67,7 +67,7 @@ export const messages = pgTable(
       .default(sql`'[]'::jsonb`),
     feedback: jsonb("feedback"), // Enhanced feedback data in JSON format (supports both legacy enum values and new structure)
     tokensUsed: integer("tokens_used").default(0), // Total tokens used for this message
-    cost: numeric("cost").default("0"), // Actual cost in dollars for this LLM call
+    cost: numeric("cost", { precision: 10, scale: 6 }).notNull().default("0"), // Actual cost in dollars for this LLM call
   },
   (table) => ({
     chatIdIndex: index("chat_id_index").on(table.chatId),
